requirements.txt
asgiref==3.5.2
attrs==23.2.0
Django==4.0.4
djangorestframework==3.14.0
drf-spectacular==0.27.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.2.0
